feat: Add LottoComputer class
The class is commonly used for comparing two lists(winner lotto,
lottos in LottoRepository). The class contains method that
calls RandomGenerator and also it saves the random number list
to LottoRepository.

Closes #33

+++ src/main/java/lotto/model/computer/LottoComputer.java
@@ -36,7 +36,7 @@ public class LottoComputer {
         LottoRandomGenerator lottoRandomGenerator = new LottoRandomGenerator();
 
         this.money = money;
-        for (int i = 0; i < money.moneyDivide1000(); i++) {
+        for (long i = 0; i < money.moneyDivide1000(); i++) {
             lottoLine = new LottoLine(sortLotto(createLottoNumber(lottoRandomGenerator)));
             lottoRepository.saveLotto(lottoLine);
         }
